Model expression use relationship should be a calculated field
The "model expression" in the past allowed the mixing of column
names produced in refSql with relationship columns, causing
confusion for users. This PR redefines the usage scenarios for expr and calc.

Currently, in the model, expr can only utilize columns from
the baseObject or those produced in refSql. Relationship columns
cannot be applied in expr. If one intends to use a relationship column,
it must be defined in calc. Additionally, calc can only utilize
columns within the model itself and cannot use columns from baseObject or refSql.

And users should be aware that model primary key and relationship join key in
join critera shouldn't use `calc`.

In metric, all expr in dimensions and measure are treated as calculated field.

This pr also remove primary key and join key validation as they should be placed
in validation rules not in runtime.

Enable to many metric test