fix(model): bigquery, avoid fetching duplicated fields when fetching it
The issue is present since the implementation of the connector.
It's trigger on our bigquery tables because we have partitionned tables.

Indeed the TABLE_OPTIONS view can have multiple entries for each field
(in our case the description and the require_partition_filter values)
thus with the left join fields were present for each entry in the
TABLE_OPTIONS.

So only join of the value that we want i.e. the description.

This was leading to breaking the import of the metadata later because
fields were duplicated.